Update challengeA_q2_1_directblockchain.py
Added notes for readability, this function assists my answer in Challenge A by verifying a claim I made when citing a statistic from the ETH Zurich research paper. This program checks to see if it is valid for us to consider Uniswap as having a relatively low participation rate. 

It returns to us proposals statistics and gives us the participation rate for recent proposals